+++ components/Monaco.vue
@@ -3,6 +3,7 @@
 </template>
 
 <script>
+import { uniq, range, union, difference } from 'lodash';
 import { storeToRefs } from 'pinia';
 import * as monaco from 'monaco-editor';
 import {
@@ -12,6 +13,7 @@ import {
     onMounted,
     useContext,
     onBeforeUnmount,
+    unref,
 } from '@nuxtjs/composition-api';
 import { useResizeObserver } from '@vueuse/core';
 import themes from 'monaco-themes/themes/themelist.json';
@@ -20,21 +22,55 @@ import usePreferencesStore from '@/composables/usePreferencesStore';
 
 export default {
     props: {
-        value: String,
-        width: Number,
-        height: Number,
-        language: String,
-        heightOffset: Number,
-        tabSize: [String, Number],
+        value: {
+            type: String,
+            default: '',
+        },
+        added: {
+            type: Array,
+            default: () => [],
+        },
+        removed: {
+            type: Array,
+            default: () => [],
+        },
+        focused: {
+            type: Array,
+            default: () => [],
+        },
+        width: {
+            type: Number,
+            default: 0,
+        },
+        height: {
+            type: Number,
+            default: 0,
+        },
+        language: {
+            type: String,
+            default: 'php',
+        },
+        heightOffset: {
+            type: Number,
+            default: 0,
+        },
+        tabSize: {
+            type: [String, Number],
+            default: 4,
+        },
     },
 
     setup(props, { emit }) {
-        const { language, tabSize, value, width, height } = toRefs(props);
+        const { language, tabSize, value, width, height, added, removed, focused } = toRefs(props);
 
         const { $bus } = useContext();
 
         const root = ref(null);
         const editor = ref(null);
+
+        const addedLineDecos = ref({});
+        const removedLineDecos = ref([]);
+        const focusedLineDecos = ref([]);
 
         const { isDarkMode } = storeToRefs(useApplicationStore());
 
@@ -49,6 +85,20 @@ export default {
             }
         };
 
+        const makeDecoration = (range, type) => {
+            return {
+                options: {
+                    stickiness: 1,
+                    marginClassName: `line ${type}`,
+                },
+                range: range,
+            };
+        };
+
+        const selectionToRange = (selection) => {
+            return range(selection.startLineNumber, selection.endLineNumber + 1);
+        };
+
         useResizeObserver(document.body, updateLayout);
 
         Object.keys(themes).forEach((theme) => {
@@ -57,17 +107,39 @@ export default {
             monaco.editor.defineTheme(theme, require(`monaco-themes/themes/${filename}.json`));
         });
 
-        watch(editorDarkTheme, (theme) => {
-            if (isDarkMode.value) {
-                monaco.editor.setTheme(theme);
-            }
-        });
+        const makeHighlightCallback = (className, ref) => () => {
+            let selected = selectionToRange(editor.value.getSelection());
 
-        watch(editorLightTheme, (theme) => {
-            if (!isDarkMode.value) {
-                monaco.editor.setTheme(theme);
+            const alreadySelected = unref(ref).filter((line) => selected.includes(line));
+
+            // If the entire selected range is already selected, we will
+            // take all of the selections and filter out the already
+            // selected lines, effectively "toggling" them off.
+            if (difference(selected, alreadySelected).length === 0) {
+                selected = unref(ref).filter((line) => !alreadySelected.includes(line));
+            } else {
+                selected = union(unref(ref), selected);
             }
-        });
+
+            emit(`update:${className}`, selected);
+        };
+
+        const makeDecosCallback = (className, ref) => (lines) => {
+            ref.value = editor.value.deltaDecorations(
+                ref.value,
+                lines.map((line) =>
+                    makeDecoration(
+                        {
+                            startColumn: 0,
+                            endColumn: 0,
+                            startLineNumber: line,
+                            endLineNumber: line,
+                        },
+                        className
+                    )
+                )
+            );
+        };
 
         onMounted(async () => {
             editor.value = monaco.editor.create(root.value, {
@@ -80,7 +152,32 @@ export default {
                 fixedOverflowWidgets: true,
                 renderLineHighlight: false,
                 scrollBeyondLastLine: false,
+                lineNumbersMinChars: 6,
                 theme: isDarkMode.value ? editorDarkTheme.value : editorLightTheme.value,
+            });
+
+            editor.value.addAction({
+                id: 'add',
+                label: 'Added Line',
+                contextMenuGroupId: 'actions',
+                keybindings: [monaco.KeyMod.CtrlCmd | monaco.KeyMod.Shift | monaco.KeyCode.KeyA],
+                run: makeHighlightCallback('added', added),
+            });
+
+            editor.value.addAction({
+                id: 'remove',
+                label: 'Removed Line',
+                contextMenuGroupId: 'actions',
+                keybindings: [monaco.KeyMod.CtrlCmd | monaco.KeyMod.Shift | monaco.KeyCode.KeyR],
+                run: makeHighlightCallback('removed', removed),
+            });
+
+            editor.value.addAction({
+                id: 'focus',
+                label: 'Focused Line',
+                contextMenuGroupId: 'actions',
+                keybindings: [monaco.KeyMod.CtrlCmd | monaco.KeyMod.Shift | monaco.KeyCode.KeyF],
+                run: makeHighlightCallback('focused', focused),
             });
 
             editor.value.onDidChangeModelContent((event) => {
@@ -111,7 +208,38 @@ export default {
                 }
             });
 
+            watch(value, () => {
+                const decos = editor.value
+                    .getModel()
+                    .getAllDecorations()
+                    .filter((deco) => deco.options.marginClassName?.includes('line'));
+
+                ['added', 'removed', 'focused'].forEach((className) => {
+                    const lines = decos
+                        .filter((deco) => deco.options.marginClassName?.includes(className))
+                        .map((deco) => deco.range.startLineNumber);
+
+                    emit(`update:${className}`, uniq(lines));
+                });
+            });
+
             watch([width, height], updateLayout);
+
+            watch(editorDarkTheme, (theme) => {
+                if (isDarkMode.value) {
+                    monaco.editor.setTheme(theme);
+                }
+            });
+
+            watch(editorLightTheme, (theme) => {
+                if (!isDarkMode.value) {
+                    monaco.editor.setTheme(theme);
+                }
+            });
+
+            watch(added, makeDecosCallback('added', addedLineDecos), { immediate: true });
+            watch(removed, makeDecosCallback('removed', removedLineDecos), { immediate: true });
+            watch(focused, makeDecosCallback('focused', focusedLineDecos), { immediate: true });
         });
 
         onBeforeUnmount(() => editor.value?.dispose());
@@ -126,18 +254,41 @@ export default {
     @apply focus:ring-0;
 }
 
+.monaco-editor .line.focused::before {
+    @apply bg-blue-500 rounded w-1.5 h-1.5 whitespace-pre ml-1;
+    content: ' ';
+}
+
+.monaco-editor .line.added::before {
+    @apply bg-green-500 rounded w-1.5 h-1.5 whitespace-pre ml-3;
+    content: ' ';
+}
+
+.monaco-editor .line.removed::before {
+    @apply bg-red-500 rounded w-1.5 h-1.5 whitespace-pre ml-5;
+    content: ' ';
+}
+
+.monaco-editor .line {
+    @apply flex items-center justify-start;
+}
+
 .monaco-editor .parameter-hints-widget {
     border: 0;
 }
+
 .monaco-editor .parameter-hints-widget .signature {
     padding: 0;
 }
+
 .monaco-editor .suggest-widget {
     border: 0;
 }
+
 .monaco-editor.vs-dark .suggest-widget {
     border: 0;
 }
+
 .monaco-editor.rename-box,
 .monaco-hover {
     top: 0;