+++ content/changelog.md
@@ -1,3 +1,10 @@
+## June 29th, 2022
+
+**Added**
+
+-   Close/open animations for tabs and preview controls
+-   Rebuilt and integrated add/remove/focus line system to into code editor (Monaco)
+
 ## June 13th, 2022
 
 **Added**